+++ src/main/java/util/blockstates/BlockHelper.java
@@ -1,10 +1,13 @@
 package util.blockstates;
 
 import java.util.HashSet;
+import java.util.LinkedHashSet;
 import java.util.LinkedList;
 import java.util.Queue;
 import java.util.Set;
 
+import net.chaos.chaosmod.world.events.tree.TreeTask;
+import net.chaos.chaosmod.world.events.tree.TreeTaskManager;
 import net.minecraft.block.Block;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.enchantment.EnchantmentHelper;
@@ -22,7 +25,7 @@ import net.minecraftforge.oredict.OreDictionary;
 public class BlockHelper {
 
 	/**
-	 * destroys the origin block and all the surroundings
+	 * destroys the origin block and all the surroundings instantly (for progressive see method below)
 	 * @param world
 	 * @param origin
 	 * @param targetState origin blockstate
@@ -53,9 +56,10 @@ public class BlockHelper {
 				block.onBlockDestroyedByPlayer(world, pos, state);
 				block.harvestBlock(world, player, pos, state, world.getTileEntity(pos), stack);
 
+				int silk_touch = EnchantmentHelper.getEnchantmentLevel(Enchantments.SILK_TOUCH, stack);
 				int fortune = EnchantmentHelper.getEnchantmentLevel(Enchantments.FORTUNE, stack);
 				int xp = block.getExpDrop(state, world, pos, fortune);
-				if (xp > 0)
+				if (xp > 0 && silk_touch == 0)
 					block.dropXpOnBlockBreak(world, pos, xp);
 
 				world.setBlockToAir(pos);
@@ -71,6 +75,45 @@ public class BlockHelper {
 				}
 			}
 		}
+	}
+
+	public static void progressiveDestroyConnectedBlocks(World world, BlockPos origin, IBlockState targetState, EntityPlayer player,
+			ItemStack stack) {
+		Queue<BlockPos> toCheck = new LinkedList<>();
+		Set<BlockPos> visited = new HashSet<>();
+		Set<BlockPos> ores = new LinkedHashSet<>();
+
+		toCheck.add(origin);
+
+		while (!toCheck.isEmpty()) {
+			BlockPos pos = toCheck.poll();
+
+			if (visited.contains(pos))
+				continue;
+
+			visited.add(pos);
+
+			IBlockState state = world.getBlockState(pos);
+			Block current = state.getBlock();
+			Block target = targetState.getBlock();
+
+			if (current == target) {
+				ores.add(pos);
+
+				for (EnumFacing facing : EnumFacing.values()) {
+					BlockPos neighbor = pos.offset(facing);
+					if (!visited.contains(neighbor)) {
+						Block neighborBlock = world.getBlockState(neighbor).getBlock();
+						if (neighborBlock == target) {
+							toCheck.add(neighbor);
+						}
+					}
+				}
+			}
+		}
+
+		TreeTask task = new TreeTask(new LinkedList<>(ores), player, stack);
+		TreeTaskManager.addTask(world, task);
 	}
 
 	public static boolean isOreBlock(IBlockState state) {